Adição de novos atributos
